Separate utility functions, so the pdf generation can run on the client
